Add support for the patch for PHP bug 76153
The upstream patch for PHP bug 76153 adds several lines to acinclude.m4
so that additional configuration is passed to ICU during compilation.

This commit adds support for that patch by adding a specific diff for
PHP 7.1+. It also takes into account the variable renaming that took
place a week later.